Update PHPUnit to version 10 and add the laminas coding standard
I don't know if the latest version of PHPUnit is required, but I figured
that as part of integrating the laminas coding standard (in place of the
outdated Zend coding standard) that it would make sense to do so, as
part of the process. Any related dependencies are also updated in this
change.